extracted FileStore to separate file
