+++ config.example.js
@@ -1,7 +1,7 @@
 module.exports = {
   domain : 'localhost:5000',
   listenOn: '5000',
-  profilesDir: 'data',
+  dataDir: 'data',
   audiences: ['http://localhost:3000'],
   secrets: {
     cookie: "BikOcikteuvMyecCergEsyidTydwapal",
